Move Rust code into its own sub-directory
This moves the Rust code into its own directory so I can keep it separate from the Cypress code.

I also split up the `.gitignore` contents so that there's a Rust-specific set of ignore directives in the new `rust` directory and just the VSCode directives are in the top level `.gitignore`.