Attempt to write unit tests
I'm rather stuck because I think I have to mock the component, and I can't seem to get `mock!` to work the way I think it maybe should work?

The other option would be to separate the state and the statement manipulation out of the component so I'm testing my state object instead of the component. I kind of feel like this would be separating the Model (in MVC) from … the control? I'm not sure.